Added LRC file supports + some code refactoring
Added LRC file supports + some code refactoring
