+++ src/Notepads/Controls/TextEditor/TextEditorCore.cs
@@ -10,7 +10,6 @@
     using Windows.ApplicationModel.DataTransfer;
     using Windows.Foundation;
     using Windows.System;
-    using Windows.UI;
     using Windows.UI.Core;
     using Windows.UI.Text;
     using Windows.UI.Xaml;
@@ -138,25 +137,16 @@
             LostFocus += OnLostFocus;
             Loaded += OnLoaded;
 
-            EditorSettingsService.OnFontFamilyChanged += EditorSettingsService_OnFontFamilyChanged;
-            EditorSettingsService.OnFontSizeChanged += EditorSettingsService_OnFontSizeChanged;
-            EditorSettingsService.OnFontStyleChanged += EditorSettingsService_OnFontStyleChanged;
-            EditorSettingsService.OnFontWeightChanged += EditorSettingsService_OnFontWeightChanged;
-            EditorSettingsService.OnDefaultTextWrappingChanged += EditorSettingsService_OnDefaultTextWrappingChanged;
-            EditorSettingsService.OnHighlightMisspelledWordsChanged += EditorSettingsService_OnHighlightMisspelledWordsChanged;
-            EditorSettingsService.OnDefaultDisplayLineNumbersViewStateChanged += EditorSettingsService_OnDefaultDisplayLineNumbersViewStateChanged;
-            EditorSettingsService.OnDefaultLineHighlighterViewStateChanged += EditorSettingsService_OnDefaultLineHighlighterViewStateChanged;
-
             SelectionChanged += OnSelectionChanged;
             TextWrappingChanged += OnTextWrappingChanged;
             SizeChanged += OnSizeChanged;
             FontSizeChanged += OnFontSizeChanged;
 
-            ThemeSettingsService.OnAccentColorChanged += ThemeSettingsService_OnAccentColorChanged;
-
             // Init shortcuts
             _keyboardCommandHandler = GetKeyboardCommandHandler();
             _mouseCommandHandler = GetMouseCommandHandler();
+
+            HookExternalEvents();
 
             Window.Current.CoreWindow.Activated += OnCoreWindowActivated;
         }
@@ -222,21 +212,12 @@
             _renderedLineNumberBlocks.Clear();
             _miniRequisiteIntegerTextRenderingWidthCache.Clear();
 
-            EditorSettingsService.OnFontFamilyChanged -= EditorSettingsService_OnFontFamilyChanged;
-            EditorSettingsService.OnFontSizeChanged -= EditorSettingsService_OnFontSizeChanged;
-            EditorSettingsService.OnFontStyleChanged -= EditorSettingsService_OnFontStyleChanged;
-            EditorSettingsService.OnFontWeightChanged -= EditorSettingsService_OnFontWeightChanged;
-            EditorSettingsService.OnDefaultTextWrappingChanged -= EditorSettingsService_OnDefaultTextWrappingChanged;
-            EditorSettingsService.OnHighlightMisspelledWordsChanged -= EditorSettingsService_OnHighlightMisspelledWordsChanged;
-            EditorSettingsService.OnDefaultDisplayLineNumbersViewStateChanged -= EditorSettingsService_OnDefaultDisplayLineNumbersViewStateChanged;
-            EditorSettingsService.OnDefaultLineHighlighterViewStateChanged -= EditorSettingsService_OnDefaultLineHighlighterViewStateChanged;
-
             SelectionChanged -= OnSelectionChanged;
             TextWrappingChanged -= OnTextWrappingChanged;
             SizeChanged -= OnSizeChanged;
             FontSizeChanged -= OnFontSizeChanged;
 
-            ThemeSettingsService.OnAccentColorChanged -= ThemeSettingsService_OnAccentColorChanged;
+            UnhookExternalEvents();
 
             Window.Current.CoreWindow.Activated -= OnCoreWindowActivated;
         }
@@ -295,80 +276,6 @@
                 new MouseCommand<PointerRoutedEventArgs>(true, false, false, false, false, false, ChangeZoomingBasedOnMouseInput),
                 new MouseCommand<PointerRoutedEventArgs>(true, false, false, OnPointerLeftButtonDown),
             }, this);
-        }
-
-        private async void EditorSettingsService_OnFontFamilyChanged(object sender, string fontFamily)
-        {
-            await Dispatcher.CallOnUIThreadAsync(() =>
-            {
-                FontFamily = new FontFamily(fontFamily);
-                SetDefaultTabStopAndLineSpacing(FontFamily, FontSize);
-            });
-        }
-
-        private async void EditorSettingsService_OnFontSizeChanged(object sender, int fontSize)
-        {
-            await Dispatcher.CallOnUIThreadAsync(() =>
-            {
-                FontSize = fontSize;
-            });
-        }
-
-        private async void EditorSettingsService_OnFontStyleChanged(object sender, FontStyle fontStyle)
-        {
-            await Dispatcher.CallOnUIThreadAsync(() =>
-            {
-                FontStyle = fontStyle;
-            });
-        }
-
-        private async void EditorSettingsService_OnFontWeightChanged(object sender, FontWeight fontWeight)
-        {
-            await Dispatcher.CallOnUIThreadAsync(() =>
-            {
-                FontWeight = fontWeight;
-            });
-        }
-
-        private async void EditorSettingsService_OnDefaultTextWrappingChanged(object sender, TextWrapping textWrapping)
-        {
-            await Dispatcher.CallOnUIThreadAsync(() =>
-            {
-                TextWrapping = textWrapping;
-            });
-        }
-
-        private async void EditorSettingsService_OnHighlightMisspelledWordsChanged(object sender, bool isSpellCheckEnabled)
-        {
-            await Dispatcher.CallOnUIThreadAsync(() =>
-            {
-                IsSpellCheckEnabled = isSpellCheckEnabled;
-            });
-        }
-
-        private async void EditorSettingsService_OnDefaultDisplayLineNumbersViewStateChanged(object sender, bool displayLineNumbers)
-        {
-            await Dispatcher.CallOnUIThreadAsync(() =>
-            {
-                DisplayLineNumbers = displayLineNumbers;
-            });
-        }
-
-        private async void EditorSettingsService_OnDefaultLineHighlighterViewStateChanged(object sender, bool displayLineHighlighter)
-        {
-            await Dispatcher.CallOnUIThreadAsync(() =>
-            {
-                DisplayLineHighlighter = displayLineHighlighter;
-            });
-        }
-
-        private async void ThemeSettingsService_OnAccentColorChanged(object sender, Color color)
-        {
-            await Dispatcher.CallOnUIThreadAsync(() =>
-            {
-                SelectionHighlightColor = new SolidColorBrush(color);
-                SelectionHighlightColorWhenNotFocused = new SolidColorBrush(color);
-            });
         }
 
         private void OnCoreWindowActivated(CoreWindow sender, WindowActivatedEventArgs args)
