+++ src/Notepads/Strings/cs-CZ/Manifest.resw
@@ -221,6 +221,10 @@
     <value>Logovací soubor</value>
     <comment>.log</comment>
   </data>
+  <data name="LrcFileDisplayName" xml:space="preserve">
+    <value>Lyrics File</value>
+    <comment>.lrc</comment>
+  </data>
   <data name="LuaFileDisplayName" xml:space="preserve">
     <value>Soubor LUA</value>
     <comment>.lua</comment>
