+++ src/Notepads/Strings/or-IN/Manifest.resw
@@ -221,6 +221,10 @@
     <value>ଲଗ୍ ଫାଇଲ୍</value>
     <comment>.log</comment>
   </data>
+  <data name="LrcFileDisplayName" xml:space="preserve">
+    <value>Lyrics File</value>
+    <comment>.lrc</comment>
+  </data>
   <data name="LuaFileDisplayName" xml:space="preserve">
     <value>ଲୁଆ ଫାଇଲ୍</value>
     <comment>.lua</comment>
