+++ src/Notepads/Strings/pt-BR/Manifest.resw
@@ -221,6 +221,10 @@
     <value>Arquivo Log</value>
     <comment>.log</comment>
   </data>
+  <data name="LrcFileDisplayName" xml:space="preserve">
+    <value>Lyrics File</value>
+    <comment>.lrc</comment>
+  </data>
   <data name="LuaFileDisplayName" xml:space="preserve">
     <value>Arquivo LUA</value>
     <comment>.lua</comment>
