+++ src/Notepads/Strings/tr-TR/Manifest.resw
@@ -221,6 +221,10 @@
     <value>Log File</value>
     <comment>.log</comment>
   </data>
+  <data name="LrcFileDisplayName" xml:space="preserve">
+    <value>Lyrics File</value>
+    <comment>.lrc</comment>
+  </data>
   <data name="LuaFileDisplayName" xml:space="preserve">
     <value>LUA File</value>
     <comment>.lua</comment>
