+++ src/Notepads/Strings/zh-TW/Manifest.resw
@@ -221,6 +221,10 @@
     <value>日誌檔</value>
     <comment>.log</comment>
   </data>
+  <data name="LrcFileDisplayName" xml:space="preserve">
+    <value>Lyrics File</value>
+    <comment>.lrc</comment>
+  </data>
   <data name="LuaFileDisplayName" xml:space="preserve">
     <value>LUA 檔案</value>
     <comment>.lua</comment>
